+++ apps/web/src/app/docs/components/button/examples/usage.tsx
@@ -1,9 +1,5 @@
 import { Button } from "@jamsrui/react";
 
 export const ButtonUsage = () => {
-  return (
-    <Button>
-      Button
-    </Button>
-  );
+  return <Button>Button</Button>;
 };